Se agrega cuadro de confirmacion para cuando se elimina una pelicula

+++ webapp/WEB-INF/views/peliculas/listPeliculas.jsp
@@ -66,7 +66,7 @@
 	                    <a href="${ urlEdit }/${ pelicula.id }" class="btn btn-success btn-sm" role="button" title="Edit" >
 	                    	<span class="glyphicon glyphicon-pencil"></span>
                     	</a>
-	                    <a href="${ urlDelete }/${ pelicula.id }" class="btn btn-danger btn-sm" role="button" title="Eliminar" >
+	                    <a href="${ urlDelete }/${ pelicula.id }" onclick="return confirm('¿Esta seguro?');" class="btn btn-danger btn-sm" role="button" title="Eliminar" >
 	                    	<span class="glyphicon glyphicon-trash"></span>
                     	</a>
 	                </td>
